restructure header social links
