+++ src/components/headerBody/index.js
@@ -4,6 +4,7 @@ import Contributors from "../contributors"
 import Github from "../github"
 import Twitter from "../twitter"
 import Slack from "../slack"
+import Discourse from "../discourse"
 import ContributorGuest from "../contributorGuest"
 import './style.css';
 
@@ -31,19 +32,18 @@ const HeaderBody = ({ title, subtitle, description, slug, contributors, featured
           <time dateTime={isoDate} pubdate="pubdate">{lastReviewed}</time>
         </p>
         {!featured && <Contributors contributors={contributors} />}
-
+        <Discourse
+          pageTitle={title}
+          path={slug}
+        />
+        <Slack />
+        {featured && <ContributorGuest contributor={contributor} />}
+        <br/><br/>
         <Github
           pageTitle={title}
           path={slug}
           editPath={editPath}
         />
-        <Twitter
-          pageTitle={title}
-          path={slug}
-        />
-        <Slack />
-
-        {featured && <ContributorGuest contributor={contributor} />}
       </header>
     </>
   )
